+++ R/serviceName.R
@@ -2,46 +2,9 @@
 # Contributed by John Mount jmount@win-vector.com , ownership assigned to Win-Vector LLC.
 # Win-Vector LLC currently distributes this code without intellectual property indemnification, warranty, claim of fitness of purpose, or any other guarantee under a GPL3 license.
 
-#' Return a cannonical name of the data service hosting a given data object.
-#'
-#'
-#' @param df data object
-#' @param expectedNames some names to canonicalize to.
-#' @return cannonical service name (lenght 1 character array)
-#'
-#' @examples
-#'
-#' replyr_dataServiceName(data.frame(x=1))
-#' replyr_dataServiceName(dplyr::as.tbl(data.frame(x=1)))
-#'
-#' @export
-replyr_dataServiceName <- function(df,
-  expectedNames=c('src_sqlite','spark_connection','src_spark',
-                  'src_mysql', 'src_postgres')) {
-  cls <- class(df)
-  if(length(cls)<=1) {
-    return(cls)  # "data.frame" case
-  }
-  if(('tbl' %in% cls) && all(cls %in% c("data.frame","tbl","tbl_df"))) {
-    return("tbl")
-  }
-  srvcls <- sort(class(df$src))
-  if(is.null(srvcls)) {
-    # give up
-    return(paste(sort(cls),collapse=' '))
-  }
-  for(m in expectedNames) {
-    if(m %in% srvcls) {
-      return(m)
-    }
-  }
-  # give up
-  return(paste(sort(srvcls),collapse=' '))
-}
 
 #' Get the "remote data source" where a data.frame like object lives.
 #'
-#' NOT TESTED YET!
 #'
 #' @param df data.frame style object
 #' @return source (string if data.frame, tlb, or data.table, NULL if unknown, remote source otherwise)
@@ -77,3 +40,79 @@ replyr_get_src <- function(df) {
   # unknown
   return(NULL)
 }
+
+#' Test if data is local.
+#'
+#' @param d data frame
+#' @return TRUE if local data (data.frame, tbl/tibble)
+#'
+#' @examples
+#'
+#' replyr_is_local_data(data.frame(x=1:3))
+#'
+#' @export
+replyr_is_local_data <- function(d) {
+  if("NULL" %in% class(d)) {
+    return(TRUE)
+  }
+  sc <- replyr_get_src(d)
+  if(is.null(sc) || is.character(sc)) {
+    return(TRUE)
+  }
+  return(FALSE)
+}
+
+#' Test if data is Spark.
+#'
+#' @param d data frame
+#' @return TRUE if Spark data
+#'
+#' @examples
+#'
+#' replyr_is_Spark_data(data.frame(x=1:3))
+#'
+#' @export
+replyr_is_Spark_data <- function(d) {
+  if("NULL" %in% class(d)) {
+    return(FALSE)
+  }
+  sc <- replyr_get_src(d)
+  if(is.null(sc) || is.character(sc)) {
+    return(FALSE)
+  }
+  if(length(grep('spark_', tolower(class(sc$con))))>0) {
+    return(TRUE)
+  }
+  if(length(grep('spark_', tolower(class(sc))))>0) {
+    return(TRUE)
+  }
+  return(FALSE)
+}
+
+#' Test if data is MySQL.
+#'
+#' @param d data frame
+#' @return TRUE if Spark data
+#'
+#' @examples
+#'
+#' replyr_is_MySQL_data(data.frame(x=1:3))
+#'
+#' @export
+replyr_is_MySQL_data <- function(d) {
+  if("NULL" %in% class(d)) {
+    return(FALSE)
+  }
+  sc <- replyr_get_src(d)
+  if(is.null(sc) || is.character(sc)) {
+    return(FALSE)
+  }
+  if(length(grep('mysql', tolower(class(sc$con))))>0) {
+    return(TRUE)
+  }
+  if(length(grep('mysql', tolower(class(sc))))>0) {
+    return(TRUE)
+  }
+  return(FALSE)
+}
+